worker/requirements.txt
#adminchat/requirements.txt
django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
python-decouple==3.8
django-filter==23.1
django-cors-headers==3.14.0
django-storages==1.13.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Pillow
drf-yasg==1.21.7
requests
boto3
PyPDF2
openpyxl
python-docx 
pgvector
psycopg2-binary
numpy>=1.20.0           # Dependencia de pgvector
celery==5.3.6
redis==4.6.0

django-celery-results==2.5.1
whitenoise==6.6.0
